+++ pom.xml
@@ -7,11 +7,13 @@
   <version>1.0-SNAPSHOT</version>
   <name>semi Maven Webapp</name>
   <url>http://maven.apache.org</url>
+
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <maven.compiler.source>17</maven.compiler.source>
     <maven.compiler.target>17</maven.compiler.target>
   </properties>
+
   <dependencies>
     <dependency>
       <groupId>org.apache.tomcat.embed</groupId>
@@ -24,6 +26,7 @@
       <artifactId>tomcat-embed-websocket</artifactId>
       <version>9.0.83</version>
     </dependency>
+
     <dependency>
       <groupId>javax.servlet</groupId>
       <artifactId>jstl</artifactId>
