+++ src/main/resources/mybatis-config.xml
@@ -51,6 +51,6 @@
 
     <mappers>
         <mapper resource="mapper/member/member-mapper.xml"/>
-        <mapper resource="mapper/member/semitest-mapper.xml"/>
+        <mapper resource="mapper/qnaBoard/qnaBoard-mapper.xml"/>
     </mappers>
 </configuration>
